Update rest of components
