+++ frontend/components/Header.tsx
@@ -1,0 +1,11 @@
+import Head from 'next/head';
+
+const Header = () => (
+  <Head>
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta charSet="utf-8" />
+    <title>WordPress + React Starter Kit Frontend by Postlight</title>
+  </Head>
+);
+
+export default Header;
