+++ frontend/components/Layout.tsx
@@ -1,6 +1,4 @@
-import React from 'react';
 import Header from './Header';
-import Menu from './Menu';
 import Footer from './Footer';
 
 const Layout = props => {
