+++ frontend/pages/index.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 import Link from 'next/link';
 import Router from 'next/router';
 import WPAPI from 'wpapi';
